backend/requirements.txt
Flask
openai-whisper
requests
TTS
torch